Move all routes to a routes folder
