+++ lib/db/dbPool.js
@@ -1,4 +1,4 @@
-var logger = require('./logger');
+var logger = require('../logger');
 var nconf  = require('nconf');
 var fs     = require('fs');
 var ini    = require('ini');
