+++ lib/db/event.js
