+++ lib/db/feature.js
@@ -1,8 +1,8 @@
-var eventStore      = require('./eventStore');
-var eventType       = require('./eventType');
-var logger          = require('./logger');
+var eventStore      = require('../eventStore');
+var eventType       = require('../eventType');
+var logger          = require('../logger');
 var knex            = require('./dbPool');
-var NotFoundError   = require('./error/NotFoundError');
+var NotFoundError   = require('../error/NotFoundError');
 var FEATURE_COLUMNS = ['name', 'description', 'enabled', 'strategy_name', 'parameters'];
 
 eventStore.on(eventType.featureCreated, function (event) {
