+++ lib/db/strategy.js
@@ -1,8 +1,8 @@
-var eventStore      = require('./eventStore');
-var eventType       = require('./eventType');
-var logger          = require('./logger');
+var eventStore      = require('../eventStore');
+var eventType       = require('../eventType');
+var logger          = require('../logger');
 var knex            = require('./dbPool');
-var NotFoundError   = require('./error/NotFoundError');
+var NotFoundError   = require('../error/NotFoundError');
 var STRATEGY_COLUMNS = ['name', 'description', 'parameters_template'];
 
 eventStore.on(eventType.strategyCreated, function (event) {
