+++ lib/eventStore.js
@@ -1,5 +1,5 @@
 var util = require('util'),
-    eventDb = require('./eventDb'),
+    eventDb = require('./db/event'),
     EventEmitter = require('events').EventEmitter;
 
 function EventStore() {
