+++ lib/routes/event.js
@@ -1,5 +1,5 @@
-var eventDb     = require('./eventDb');
-var eventDiffer = require('./eventDiffer');
+var eventDb     = require('../db/event');
+var eventDiffer = require('../eventDiffer');
 
 module.exports = function (app) {
     app.get('/events', function (req, res) {
