+++ lib/routes/feature-archive.js
@@ -1,9 +1,9 @@
-var logger              = require('./logger');
-var eventStore          = require('./eventStore');
-var eventType           = require('./eventType');
-var featureDb           = require('./featureDb');
-var ValidationError     = require('./error/ValidationError');
-var validateRequest     = require('./error/validateRequest');
+var logger              = require('../logger');
+var eventStore          = require('../eventStore');
+var eventType           = require('../eventType');
+var featureDb           = require('../db/feature');
+var ValidationError     = require('../error/ValidationError');
+var validateRequest     = require('../error/validateRequest');
 
 module.exports = function (app) {
     app.get('/archive/features', function (req, res) {
