+++ lib/routes/feature.js
@@ -1,13 +1,13 @@
 var Promise             = require("bluebird");
-var logger              = require('./logger');
-var eventStore          = require('./eventStore');
-var eventType           = require('./eventType');
-var featureDb           = require('./featureDb');
-var NameExistsError     = require('./error/NameExistsError');
-var NotFoundError       = require('./error/NotFoundError');
-var ValidationError     = require('./error/ValidationError');
-var validateRequest     = require('./error/validateRequest');
-var extractUser         = require('./extractUser');
+var logger              = require('../logger');
+var eventStore          = require('../eventStore');
+var eventType           = require('../eventType');
+var featureDb           = require('../db/feature');
+var NameExistsError     = require('../error/NameExistsError');
+var NotFoundError       = require('../error/NotFoundError');
+var ValidationError     = require('../error/ValidationError');
+var validateRequest     = require('../error/validateRequest');
+var extractUser         = require('../extractUser');
 
 module.exports = function (app) {
     app.get('/features', function (req, res) {
