+++ lib/routes/health-check.js
@@ -1,5 +1,5 @@
-var knex = require('./dbPool');
-var logger = require('./logger');
+var knex = require('../db/dbPool');
+var logger = require('../logger');
 
 module.exports = function (app) {
     app.get('/health', function (req, res) {
