+++ lib/routes/strategy.js
@@ -1,13 +1,13 @@
 var Promise             = require("bluebird");
-var eventStore          = require('./eventStore');
-var eventType           = require('./eventType');
-var strategyDb          = require('./strategyDb');
-var logger              = require('./logger');
-var NameExistsError     = require('./error/NameExistsError');
-var ValidationError     = require('./error/ValidationError');
-var NotFoundError       = require('./error/NotFoundError');
-var validateRequest     = require('./error/validateRequest');
-var extractUser         = require('./extractUser');
+var eventStore          = require('../eventStore');
+var eventType           = require('../eventType');
+var strategyDb          = require('../db/strategy');
+var logger              = require('../logger');
+var NameExistsError     = require('../error/NameExistsError');
+var ValidationError     = require('../error/ValidationError');
+var NotFoundError       = require('../error/NotFoundError');
+var validateRequest     = require('../error/validateRequest');
+var extractUser         = require('../extractUser');
 
 module.exports = function (app) {
     app.get('/strategies', function (req, res) {
